@@ -9,26 +9,31 @@
     "@mui/material": "^5.9.0",
     "@testing-library/jest-dom": "^5.16.4",
     "@types/react-image-gallery": "1.0.5",
+    "bcryptjs": "^2.4.3",
     "cloudinary": "^1.30.0",
     "connect-multiparty": "^2.2.0",
     "express": "^4.18.2",
     "express-session": "^1.17.3",
     "mongoose": "^6.4.4",
     "mui": "^0.0.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-icons": "^4.4.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.7.1",
     "react-image-gallery": "^1.2.8",
     "react-router": "^5.2.0",
-    "react-router-dom": "^5.3.3",
-    "react-scripts": "^2.1.3",
+    "react-router-dom": "^5.2.0",
+    "react-scripts": "^5.0.1",
+    "react-uid": "^2.3.2",
+    "rsuite": "^5.28.1",
+    "validator": "^13.9.0",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
     "setup": "npm install && cd client && npm install && cd ..",
-    "build-run": "cd client && npm run build && cd .. && node server.js",
-    "build": "cd client && npm install && npm run build",
+    "build-run": "npm run build && node server.js",
+    "build": "react-scripts build",
     "start": "node server.js",
+    "front-start": "react-scripts start",
     "heroku-postbuild": "cd client && npm install && npm run build",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
