@@ -3,7 +3,7 @@ import React from 'react';
 import './index.css';
 
 import { withRouter } from 'react-router' 
-import Portal_list from '../Portal_list';
+import PortalList from '../Portal_list';
 import Footer from '../footer';
 import demo_img1 from '../../content-img/demo-img1.png'
 import demo_img2 from '../../content-img/demo-img2.png'
@@ -11,6 +11,8 @@ import demo_img3 from '../../content-img/demo-img3.png'
 import demo_img4 from '../../content-img/demo-img4.png'
 import demo_img5 from '../../content-img/demo-img5.png'
 import demo_img6 from '../../content-img/demo-img6.png'
+import demo_img7 from '../../content-img/demo-img7.png'
+import demo_img8 from '../../content-img/demo-img8.png'
 
 class Demo_Page extends React.Component {
    
@@ -18,24 +20,29 @@ class Demo_Page extends React.Component {
     render(){
         return (
             <div className='introduction'>
-                <Portal_list/>
+                <PortalList/>
                 <div className='demo-space'></div>
                 <div className='demo-container'>
+                    <div className='grid-img'>
+                        <img src={demo_img1} alt='demo-img1'/>
+                    </div>
                     <div className='demo-content'>
                         <p className='demo-subtitle'>界面交互，告诉 Onion 研究梗概</p>
                         <p className='demo-text'>Onion AI 分析研究者拟定研究领域的指南、共识、文献，提出有价值的研究课题建议，延展研究者的研究创新；并帮助研究者聚焦于具有临床实践意义的课题方向</p>
+                        <br/>
+                        <p className='demo-subtitle double-space'>Onion 激发灵感</p>
+                        <p className='demo-text double-space'>Onion 从多个维度提出突破“先验研究”的建议，使研究者的研究成为“迈出前沿的一小步、推进前沿的一大步”</p>
                     </div>
-                    <div className='demo-img'>
-                        <img src={demo_img1} alt='demo-img1'/>
-                    </div>
+                    
                 </div>
-                <div className='demo-container'>
+                <div className='vertical-container'>
                     <div className='grid-img'>
                         <img src={demo_img2} alt='demo-img2'/>
                     </div>
                     <div className='demo-content'>
-                        <p className='demo-subtitle'>Onion 激发灵感</p>
-                        <p className='demo-text'>Onion 从多个维度提出突破“先验研究”的建议，使研究者的研究成为“迈出前沿的一小步、推进前沿的一大步”</p>
+                        <p className='demo-subtitle'>文献积累，Onion 辅助研究者</p>
+                        <p className='demo-text'>检索的文献如果不能进行有效的分析则很难为研究提供帮助。文献积累应该怎样做？Onion 向导研究者有导向性的积累文献，并作出分析图表；</p>
+                        <p className='demo-text'>先验研究的重要信息，一目了然！</p>
                     </div>                   
                 </div>
                 <div className='demo-container'>
@@ -49,7 +56,7 @@ class Demo_Page extends React.Component {
                     </div>
                 </div>
                 <div className='demo-container'>
-                    <div className='grid-img'>
+                    <div className='grid-img4'>
                         <img src={demo_img4} alt='demo-img4'/>
                     </div>
                     <div className='demo-content'>
@@ -69,19 +76,39 @@ class Demo_Page extends React.Component {
                         <br/>
                         <br/>
                         <br/>
+                        
+                    </div>                   
+                </div>
+                <div className='vertical-container'>
+                    <div className='grid-img'>
+                        <img src={demo_img6} alt='demo-img5'/>
+                    </div>
+                    <div className='vertical-content'>
                         <p className='demo-subtitle'>确定研究“技术路径”（path）</p>
                         <p className='demo-text'>Onion 发出提示，让研究者自由选择研究技术路径（path）</p>
-                    </div>                   
+                    </div>
+                        
+                </div>
+                <div className='demo-container'>
+                    <div className='grid-img'>
+                        <img src={demo_img7} alt='demo-img7'/>
+                    </div>
+                    <div className='demo-content'>
+                        <p className='demo-subtitle'>AI 辅助，形成研究的关键框架（Framework）：</p>
+                        <p className='demo-text'>Onion 的AI算法，根据用户交互的信息，一键式生成了研究“Framework”（关键框架），在这个“Framework”里面包括了研究非常关键的、也是经常困扰临床研究者的问题之解决方案，包括：</p>
+                        <p className='demo-text'>如何假设、研究设计、是否“巢氏”、“如何随机”、“样本量估计方法”、“统计学分析方法”、“变量处理”、“偏倚控制策略”、“敏感性分析”、“亚组分析”……</p>
+                    </div>
+                    
                 </div>
                 <div className='demo-container'>
                     <div className='demo-content'>
                         <p className='demo-subtitle'>研究方案 & 报告，Onion 向导下轻松撰写</p>
                         <p className='demo-text'>规范的研究方案“模版”，针对每个模块 Onion 都发出提示，研究者只需跟随 Onion 的向导即可高效的完成一份研究方案 & 报告；</p>
                         <p className='demo-text'>质量有保证的方案、考虑全面的方案……</p>
-                        <p>以研究方案规范形式呈现给研究者，事倍功半！</p>
+                        <p className='demo-text'>以研究方案规范形式呈现给研究者，事倍功半！</p>
                     </div>
                     <div className='demo-img demo-large-img'>
-                        <img src={demo_img6} alt='demo-img6'/>
+                        <img src={demo_img8} alt='demo-img8'/>
                     </div>
                 </div>
                 <div className='demo-space'></div>
